Split out hashRouting into separate module
